chore: delete and inline babel-plugin
