--- tooling/babel-plugin/.babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "node": 8
-        },
-        "loose": true
-      }
-    ]
-  ]
-}
