--- tooling/babel-plugin/.eslintrc
@@ -1,27 +1,0 @@
-{
-  "root": true,
-  "env": {
-    "browser": true,
-    "node": true
-  },
-  "parser": "babel-eslint",
-  "parserOptions": {
-    "ecmaFeatures": {
-      "jsx": true
-    },
-    "ecmaVersion": 2018,
-    "sourceType": "module",
-    "allowImportExportEverywhere": true
-  },
-  "extends": ["plugin:react/recommended", "plugin:prettier/recommended"],
-  "plugins": ["react"],
-  "rules": {
-    "react/prop-types": "off",
-    "prettier/prettier": [
-      "error",
-      {
-        "endOfLine": "auto"
-      }
-    ]
-  }
-}
