--- tooling/babel-plugin/CHANGELOG.md
@@ -1,131 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.0.0 (2020-11-13)
-
-**Note:** Version bump only for package @chakra-ui/babel-plugin
-
-
-
-
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.0.0-rc.8 (2020-10-29)
-
-### Bug Fixes
-
-- **toast:** allow custom render in update
-  ([eb8bff9](https://github.com/chakra-ui/chakra-ui/commit/eb8bff911e6ec9de0165ab1e8f5ca10d5e022459)),
-  closes [#2362](https://github.com/chakra-ui/chakra-ui/issues/2362)
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.0.0-rc.7 (2020-10-25)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# 1.0.0-rc.6 (2020-10-25)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# 1.0.0-rc.5 (2020-09-27)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# 1.0.0-rc.4 (2020-09-25)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# 1.0.0-rc.3 (2020-08-30)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.0.0-rc.2 (2020-08-09)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [1.0.0-rc.1](https://github.com/chakra-ui/chakra-ui/compare/babel-plugin-chakra-ui@1.0.0-rc.0...babel-plugin-chakra-ui@1.0.0-rc.1) (2020-08-06)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [1.0.0-rc.0](https://github.com/chakra-ui/chakra-ui/compare/babel-plugin-chakra-ui@1.0.0-next.7...babel-plugin-chakra-ui@1.0.0-rc.0) (2020-07-26)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [1.0.0-next.7](https://github.com/chakra-ui/chakra-ui/compare/babel-plugin-chakra-ui@1.0.0-next.6...babel-plugin-chakra-ui@1.0.0-next.7) (2020-07-26)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [1.0.0-next.6](https://github.com/chakra-ui/chakra-ui/compare/babel-plugin-chakra-ui@1.0.0-next.5...babel-plugin-chakra-ui@1.0.0-next.6) (2020-07-15)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [1.0.0-next.5](https://github.com/chakra-ui/chakra-ui/compare/babel-plugin-chakra-ui@1.0.0-next.4...babel-plugin-chakra-ui@1.0.0-next.5) (2020-07-15)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.0.0-next.4 (2020-07-01)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.0.0-next.3 (2020-06-28)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.0.0-next.2 (2020-06-21)
-
-**Note:** Version bump only for package babel-plugin-chakra-ui
