--- tooling/babel-plugin/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@chakra-ui/babel-plugin",
-  "version": "1.0.0",
-  "description": "Mark custom forwardRef and memo as pure for tree shaking",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "dist/index.js",
-  "keywords": [
-    "chakra-ui",
-    "babel-plugin"
-  ],
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "prepublish": "yarn build",
-    "lint": "eslint --ext .js src",
-    "test": "jest --passWithNoTests",
-    "build": "babel src -d dist"
-  },
-  "dependencies": {
-    "@babel/helper-annotate-as-pure": "^7.12.10",
-    "@babel/helper-plugin-utils": "^7.10.4"
-  },
-  "peerDependencies": {
-    "@babel/core": "^7.12.10"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.10"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "testRegex": "/test/.*\\.test\\.js$",
-    "collectCoverageFrom": [
-      "src/**/*.js",
-      "!src/log.js"
-    ]
-  }
-}
